Add files via upload
This project implements a spam detection system using a pre-trained BERT model. The goal is to classify text messages or emails as either spam or ham (not spam) by leveraging the capabilities of BERT for natural language understanding.

BERT (Bidirectional Encoder Representations from Transformers) is fine-tuned on the spam classification task to achieve high accuracy and robust detection of spam messages.